requirements.txt

cloudinary==1.44.0
pathlib==1.0.1
pip==25.1.1
psycopg2-binary==2.9.10
python-dotenv==1.1.0
setuptools==65.5.0
streamlit==1.45.1
